Week progress
Input changes, makes it so a player can select their own input keys.
Added weapon pickups and switching. Stats for end game are now properly
computed. Animations have been implemented for enemies. Head shots.

+++ Assets/Scripts/Interfaces/PlayerHealthScript.cs
@@ -4,7 +4,7 @@ using System;
 
 public class PlayerHealthScript : MonoBehaviour {
     private Player gameUser;
-    private int playerHealthRemaining;
+    private float playerHealthRemaining;
     private int screenWidth;
     private int screenHeight;
     private float healthRemainingWidthLocation, healthRemainingHeightLocation;
